% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.r
\name{nm}
\alias{nm}
\title{Run the model either from file or "nmodel" object}
\usage{
nm(model, cmake_stdout = FALSE)
}
\arguments{
\item{model}{model control stream file or "nmodel" object}

\item{cmake_stdout}{Whether to output cmake build info to screen}
}
\value{
nmrun object
}
\description{
Run the model either from file or "nmodel" object
}
