% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm.r
\name{nm_file}
\alias{nm_file}
\title{Run NONMEM model specified in a file}
\usage{
nm_file(raw_model_path, cmake_stdout)
}
\arguments{
\item{raw_model_path}{model control stream file path}

\item{cmake_stdout}{Whether to output cmake build info to screen}
}
\value{
nmrun object
}
\description{
The function runs a NONMEM model given in a control stream file. It requires that
the data file is located in the same path as the control stream.
}
