% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_model.r
\name{nm_model}
\alias{nm_model}
\title{Create an "nmodel" object based on given data and code.}
\usage{
nm_model(data, code, tables = NULL)
}
\arguments{
\item{data}{data object for the model. It must contain necessary
columns for NONMEM models.}

\item{code}{NONMEM code, excluding $PROB, $DATA, and $INPUT.}

\item{tables}{list of variable arrays for table output. Each array
will be for one table. The table file's name is "tablex.tab", with
"x" enumerates from 1 to the number of arrays.}
}
\value{
An "nmodel" object that contains the data and a vector of
characters for control stream.
}
\description{
Create an "nmodel" object based on given data and code.
}
