% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_model.r
\name{print.nmodel}
\alias{print.nmodel}
\title{Print the generate control stream from "nmodel" object}
\usage{
\method{print}{nmodel}(model)
}
\description{
Print the generate control stream from "nmodel" object
}
