% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_nm.r
\name{set_nm}
\alias{set_nm}
\title{Set NONMEM & lic path}
\usage{
set_nm(p1, p2)
}
\arguments{
\item{p1}{NONMEM directory path}

\item{p2}{NONMEM license path}
}
\value{
Normalized path to NONMEM license directory
}
\description{
Set NONMEM & lic path
}
