% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_nm.r
\name{set_nm_lic}
\alias{set_nm_lic}
\title{Set NONMEM license path}
\usage{
set_nm_lic(p)
}
\arguments{
\item{p}{NONMEM license path}
}
\value{
Normalized path to NONMEM license
}
\description{
Set NONMEM license path
}
