% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nm_model.r
\name{write_nm_csv}
\alias{write_nm_csv}
\title{Write a data.frame to .csv compatible with NONMEM format
One can use R's read.csv to read a NONMEM .csv file:}
\usage{
write_nm_csv(data, file)
}
\arguments{
\item{data}{data object to be saved}

\item{file}{.csv file path}
}
\description{
read.csv(file, comment.char = ignore, na.strings = ".")
}
