% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_nm.r
\name{set_nm}
\alias{set_nm}
\title{Set NONMEM & lic path}
\usage{
set_nm(nm_path, nm_lic)
}
\arguments{
\item{nm_path}{NONMEM directory path}

\item{nm_lic}{NONMEM license path}
}
\value{
Normalized path to NONMEM license directory
}
\description{
Set NONMEM & lic path
}
